deployment/requirements.txt
pandas
numpy
joblib
pillow
sklearn
streamlit
matplotlib
seaborn
tensorflow-cpu